feat: 게임오버 화면, 둥근 모서리 육각형, 블럭 간격 최적화
- GameOverScreen 표시 수정 (ScreenManager 충돌 해결)
- 육각형 블럭 둥근 모서리 적용 (cornerRadius=0.08)
- 블럭 간격 벤치마크 매칭 (radius=0.478, hexSpacing=0)
- BuildHelper에서 빌드 전 SceneSetup 자동 실행
- 128배 소멸 규칙, splat 점성 효과, 숫자 축약 개선

+++ Assets/Scripts/Audio/AudioManager.cs
@@ -104,7 +104,7 @@ namespace HexaMerge.Audio
             }
         }
 
-        public static SFXType GetMergeSFXType(int resultValue)
+        public static SFXType GetMergeSFXType(double resultValue)
         {
             if (resultValue <= 64) return SFXType.MergeBasic;
             if (resultValue <= 512) return SFXType.MergeMid;
